requirements.txt
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiofiles>=23.2.0
asyncio-mqtt>=0.16.0
websockets>=11.0.0
rich>=13.6.0
click>=8.1.0
pyyaml>=6.0
toml>=0.10.0
jinja2>=3.1.0
markdown>=3.5.0
pygments>=2.16.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pillow>=10.0.0
opencv-python>=4.8.0
transformers>=4.35.0
torch>=2.1.0
tensorflow>=2.14.0
langchain>=0.0.350
langchain-openai>=0.0.2
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0
redis>=5.0.0
aiohttp>=3.8.0
typing-extensions>=4.8.0
colorama>=0.4.6
